moved all rest clients together
